+++ src/main/java/coordinate/domain/Common.java
@@ -1,8 +1,7 @@
 package coordinate.domain;
 
 public class Common {
-	public static final int MAXSIZE = 24;
-	public static final String BLANK = "  ";
+
 
 	public static Boolean isOdd(int i) {
 		return i % 2 != 0;
