+++ src/main/java/coordinate/view/PrintManager.java
@@ -4,6 +4,8 @@ import coordinate.domain.Common;
 import coordinate.domain.CoordinateManager;
 
 public class PrintManager {
+	public static final int MAXSIZE = 24;
+	public static final String BLANK = "  ";
 
 	public static void drawYaxis(CoordinateManager coordinateManager) {
 		for (int i = coordinateManager.size() - 1; i > 0; i--) {
@@ -14,24 +16,24 @@ public class PrintManager {
 
 	public static void drawXaxis() {
 		printXaxisBar();
-		for (int i = 0; i < Common.MAXSIZE + 1; i++) {
+		for (int i = 0; i < MAXSIZE + 1; i++) {
 			System.out.print(printxValue(i));
 		}
 	}
 
 	public static void printXaxisBar() {
-		System.out.print(Common.BLANK+"*");
-		for (int i = 0; i < Common.MAXSIZE + 1; i++) {
+		System.out.print(BLANK+"*");
+		for (int i = 0; i < MAXSIZE + 1; i++) {
 			System.out.print("───");
 		}
 		System.out.println();
-		System.out.print(Common.BLANK);
+		System.out.print(BLANK);
 	}
 
 	public static String printxValue(int i) {
 		if (!Common.isOdd(i))
-			return i + Common.BLANK;
-		return Common.BLANK;
+			return i + BLANK;
+		return BLANK;
 	}
 
 }
